Lambertian now working
There is a lot of debug code left in the commit. Need to remove soon™

Main problem was sphere_ray_hits(): we want the *positive*, smaller (aka closer to ray origin) root. Previously it just returns the left root, but that might be negative.

Also problem: the bounce offset vector was "any inside unit sphere", which isn't right. Lambertian is a uniformly distributed random vector on the *surface of* a unit sphere, located at tip of the normal.

BUG: everything disappears if camera looks directly down. I think it has something to do with the camera-view space basis calculation.

Interestingly, if sphere_ray_hits() only returns a hit when both roots are positive, all noise disappears from the image, but along with shading of the upper part of the sphere.
Print debugging on that case shows r1 is very big negative value. Either -100 or -99.999985
Why would that be the case?
A previous guess: on some bounces, the new ray has a origin just inside the sphere, due to rounding errors. But that can't be the case because r1 is so negative...?